Signing Back In
A check is run to see if a user currently signed out when their barcode is scanned in. If their barcode is listed as signed out, they are now prompted whether they would like to sign back in or not. Clicking yes updates their time back and attendance status cells.