Close background windows
Previously, if the user accidently clicked to the side of any popups, they could accidently lock the program because the program only allows barcode scans if now popups exist. After this commit, if the main screen is in front (meaning no popups are accessible or visible), the hidden popups are destroyed and a new session can be started.