ETR Entry
Added new popup that asks users to specify an estimated time of return if they are leaving campus by car and will return later in the day. The input format is two scrollable lists, one for hours and one for minutes. Hopefully this will adapt well for touch screens, but more testing is needed.